fix: Remove @deephaven/app-utils from @deephaven/dashboard-core-plugins dependency list
