+++ packages/app-utils/src/components/ConnectionBootstrap.tsx
@@ -1,12 +1,11 @@
-import React, { createContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { LoadingOverlay } from '@deephaven/components';
 import { useApi, useClient } from '@deephaven/jsapi-bootstrap';
 import type { IdeConnection } from '@deephaven/jsapi-types';
+import { ConnectionContext } from '@deephaven/jsapi-components';
 import Log from '@deephaven/log';
 
-const log = Log.module('@deephaven/jsapi-components.ConnectionBootstrap');
-
-export const ConnectionContext = createContext<IdeConnection | null>(null);
+const log = Log.module('@deephaven/app-utils.ConnectionBootstrap');
 
 export type ConnectionBootstrapProps = {
   /**
