+++ packages/app-utils/src/components/PluginsBootstrap.tsx
@@ -1,8 +1,10 @@
-import { type Plugin } from '@deephaven/plugin';
-import React, { createContext, useEffect, useState } from 'react';
-import { PluginModuleMap, loadModulePlugins } from '../plugins';
-
-export const PluginsContext = createContext<PluginModuleMap | null>(null);
+import React, { useEffect, useState } from 'react';
+import {
+  type Plugin,
+  type PluginModuleMap,
+  PluginsContext,
+} from '@deephaven/plugin';
+import { loadModulePlugins } from '../plugins';
 
 export type PluginsBootstrapProps = {
   /**
