+++ packages/dashboard-core-plugins/src/index.test.tsx
@@ -6,7 +6,8 @@ import { createMockStore } from '@deephaven/redux';
 import { dh } from '@deephaven/jsapi-shim';
 import { ApiContext } from '@deephaven/jsapi-bootstrap';
 import { type IdeConnection } from '@deephaven/jsapi-types';
-import { ConnectionContext, PluginsContext } from '@deephaven/app-utils';
+import { ConnectionContext } from '@deephaven/jsapi-components';
+import { PluginsContext } from '@deephaven/plugin';
 import {
   ChartPlugin,
   ConsolePlugin,
