+++ packages/dashboard-core-plugins/src/useHydrateGrid.ts
@@ -4,7 +4,7 @@ import {
   PanelHydrateFunction,
 } from '@deephaven/dashboard';
 import { useApi } from '@deephaven/jsapi-bootstrap';
-import { useConnection, useLoadTablePlugin } from '@deephaven/app-utils';
+import { useConnection } from '@deephaven/jsapi-components';
 import { Table } from '@deephaven/jsapi-types';
 import { IrisGridModelFactory } from '@deephaven/iris-grid';
 import {
@@ -13,6 +13,7 @@ import {
   isIrisGridPanelMetadata,
   isLegacyIrisGridPanelMetadata,
 } from './panels';
+import { useLoadTablePlugin } from './useLoadTablePlugin';
 
 export function useHydrateGrid<
   P extends DehydratedDashboardPanelProps = DehydratedDashboardPanelProps,
