+++ packages/jsapi-components/src/ConnectionContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+import type { IdeConnection } from '@deephaven/jsapi-types';
+
+export const ConnectionContext = createContext<IdeConnection | null>(null);
+
+export default ConnectionContext;
