+++ packages/jsapi-components/src/useConnection.ts
@@ -1,11 +1,11 @@
 import type { IdeConnection } from '@deephaven/jsapi-types';
 import { useContextOrThrow } from '@deephaven/react-hooks';
-import { ConnectionContext } from './ConnectionBootstrap';
+import { ConnectionContext } from './ConnectionContext';
 
 export function useConnection(): IdeConnection {
   return useContextOrThrow(
     ConnectionContext,
-    'No IdeConnection available in useConnection. Was code wrapped in ConnectionBootstrap or ConnectionContext.Provider?'
+    'No IdeConnection available in useConnection. Was code wrapped in ConnectionContext.Provider?'
   );
 }
