+++ packages/plugin/src/PluginsContext.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+import { type PluginModule } from './PluginTypes';
+
+export type PluginModuleMap = Map<string, PluginModule>;
+
+export const PluginsContext = createContext<PluginModuleMap | null>(null);
+
+export default PluginsContext;
