+++ packages/plugin/src/index.ts
@@ -1,3 +1,5 @@
+export * from './PluginsContext';
 export * from './PluginTypes';
-export * from './TablePlugin';
 export * from './PluginUtils';
+export * from './TablePlugin';
+export * from './usePlugins';
