+++ packages/redux/src/actions.ts
@@ -3,6 +3,7 @@ import type { ThunkAction } from 'redux-thunk';
 import type { CommandHistoryStorage } from '@deephaven/console';
 import type { FileStorage } from '@deephaven/file-explorer';
 import type { dh as DhType } from '@deephaven/jsapi-types';
+import type { PluginModuleMap } from '@deephaven/plugin';
 import {
   SET_PLUGINS,
   SET_USER,
@@ -15,7 +16,6 @@ import {
   SET_API,
 } from './actionTypes';
 import type {
-  DeephavenPluginModuleMap,
   RootState,
   ServerConfigValues,
   User,
@@ -135,9 +135,7 @@ export const setActiveTool: PayloadActionCreator<string> = payload => ({
   payload,
 });
 
-export const setPlugins: PayloadActionCreator<
-  DeephavenPluginModuleMap
-> = plugins => ({
+export const setPlugins: PayloadActionCreator<PluginModuleMap> = plugins => ({
   type: SET_PLUGINS,
   payload: plugins,
 });
